Refactor double data types to BigDecimal

+++ src/main/java/net/fosforito/partido/model/bill/BillDTO.java
@@ -3,6 +3,7 @@ package net.fosforito.partido.model.bill;
 import net.fosforito.partido.model.split.SplitDTO;
 
 import javax.validation.constraints.NotNull;
+import java.math.BigDecimal;
 import java.util.Date;
 import java.util.List;
 
@@ -15,13 +16,13 @@ public class BillDTO {
   private String category;
 
   @NotNull
-  private double totalAmount;
+  private BigDecimal totalAmount;
 
   @NotNull
   private Date billingDate;
 
   @NotNull
-  private double parts;
+  private BigDecimal parts;
 
   private List<SplitDTO> splits;
 
@@ -30,9 +31,9 @@ public class BillDTO {
 
   public BillDTO(@NotNull String description,
                  @NotNull String category,
-                 @NotNull double totalAmount,
+                 @NotNull BigDecimal totalAmount,
                  @NotNull Date billingDate,
-                 @NotNull double parts,
+                 @NotNull BigDecimal parts,
                  List<SplitDTO> splits) {
     this.description = description;
     this.category = category;
@@ -58,11 +59,11 @@ public class BillDTO {
     this.category = category;
   }
 
-  public double getTotalAmount() {
+  public BigDecimal getTotalAmount() {
     return totalAmount;
   }
 
-  public void setTotalAmount(double totalAmount) {
+  public void setTotalAmount(BigDecimal totalAmount) {
     this.totalAmount = totalAmount;
   }
 
@@ -74,11 +75,11 @@ public class BillDTO {
     this.billingDate = billingDate;
   }
 
-  public double getParts() {
+  public BigDecimal getParts() {
     return parts;
   }
 
-  public void setParts(double parts) {
+  public void setParts(BigDecimal parts) {
     this.parts = parts;
   }
